refactor(breach-history): extract DurationFilter + BucketFilter to _components/
Mechanical extraction only — no behavior changes. DurationField and
DurationRangeInputs move to _components/duration-filter.tsx; BucketFilter,
DurationBucket type, BUCKET_LABEL, ONE_HOUR, and ONE_DAY move to
_components/bucket-filter.tsx. breach-history-panel.tsx imports from the new
locations. All 51 characterization tests stay green.
