+++ ui-dashboard/src/components/breach-history-panel.tsx
@@ -15,7 +15,7 @@ import {
   POOL_DEVIATION_BREACHES_PAGE,
 } from "@/lib/queries";
 import { formatTimestamp, relativeTime } from "@/lib/format";
-import { formatDurationShort, parseDurationSeconds } from "@/lib/bridge-status";
+import { formatDurationShort } from "@/lib/bridge-status";
 import {
   formatDeviationPct,
   DEVIATION_BREACH_GRACE_SECONDS,
@@ -36,6 +36,13 @@ import {
   normalizeSearch,
 } from "@/lib/table-search";
 import { ENVIO_MAX_ROWS } from "@/lib/constants";
+import { DurationRangeInputs } from "@/components/_components/duration-filter";
+import {
+  BucketFilter,
+  type DurationBucket,
+  ONE_HOUR,
+  ONE_DAY,
+} from "@/components/_components/bucket-filter";
 
 interface Props {
   pool: Pool;
@@ -61,25 +68,6 @@ const START_REASON_LABELS: Record<BreachEventCategory, string> = {
   oracle_update: "Oracle moved",
   threshold_change: "Threshold change",
   unknown: "Unknown",
-};
-
-/**
- * Duration-range filter presets. Each one compiles to a Hasura
- * `_and`-able where clause. Buckets line up with how operators think
- * about breach severity (in-grace = WARN-only, 1h–1d = moderately bad,
- * >1d = really stuck).
- */
-type DurationBucket = "all" | "in_grace" | "short" | "long" | "ongoing";
-
-const ONE_HOUR = 3600;
-const ONE_DAY = 86400;
-
-const BUCKET_LABEL: Record<DurationBucket, string> = {
-  all: "All",
-  in_grace: "≤1h",
-  short: "1h – 1d",
-  long: "Over 1d",
-  ongoing: "Ongoing",
 };
 
 function whereForBucket(bucket: DurationBucket): Record<string, unknown> {
@@ -547,153 +535,6 @@ function BreachHistoryPanelInner({
           />
         )}
       </section>
-    </div>
-  );
-}
-
-function DurationRangeInputs({
-  minSeconds,
-  maxSeconds,
-  onMinCommit,
-  onMaxCommit,
-}: {
-  minSeconds: number | null;
-  maxSeconds: number | null;
-  onMinCommit: (seconds: number | null) => void;
-  onMaxCommit: (seconds: number | null) => void;
-}) {
-  return (
-    <div className="flex items-center gap-2">
-      <DurationField
-        ariaLabel="Minimum breach duration"
-        placeholder="Min."
-        committedSeconds={minSeconds}
-        onCommit={onMinCommit}
-      />
-      <span className="text-xs text-slate-600">–</span>
-      <DurationField
-        ariaLabel="Maximum breach duration"
-        placeholder="Max."
-        committedSeconds={maxSeconds}
-        onCommit={onMaxCommit}
-      />
-    </div>
-  );
-}
-
-/**
- * Single duration input. Keeps its own draft text so the parent only
- * re-renders (and re-fires the GraphQL query) on blur or Enter. An empty
- * draft commits `null` — clears the filter instead of leaving it stale.
- * A parse failure keeps the previous committed value and flags the input
- * with a red ring until the user fixes it.
- */
-function DurationField({
-  ariaLabel,
-  placeholder,
-  committedSeconds,
-  onCommit,
-}: {
-  ariaLabel: string;
-  placeholder: string;
-  committedSeconds: number | null;
-  onCommit: (seconds: number | null) => void;
-}) {
-  const [draft, setDraft] = useState(() =>
-    committedSeconds != null ? formatDurationShort(committedSeconds) : "",
-  );
-  const [invalid, setInvalid] = useState(false);
-  const commit = useCallback(() => {
-    const trimmed = draft.trim();
-    if (!trimmed) {
-      setInvalid(false);
-      onCommit(null);
-      return;
-    }
-    const parsed = parseDurationSeconds(trimmed);
-    if (parsed == null || parsed <= 0) {
-      setInvalid(true);
-      return;
-    }
-    setInvalid(false);
-    onCommit(parsed);
-  }, [draft, onCommit]);
-  return (
-    <input
-      type="text"
-      inputMode="text"
-      autoComplete="off"
-      placeholder={placeholder}
-      aria-label={ariaLabel}
-      value={draft}
-      onChange={(e) => {
-        setDraft(e.target.value);
-        // Clear the error as soon as they start fixing it — no point
-        // yelling while they type.
-        if (invalid) setInvalid(false);
-      }}
-      onBlur={commit}
-      onKeyDown={(e) => {
-        if (e.key === "Enter") {
-          e.preventDefault();
-          commit();
-        }
-      }}
-      className={
-        "w-20 rounded-lg border bg-slate-800 px-2 py-1 text-xs text-white placeholder-slate-500 focus:outline-none focus:ring-1 " +
-        (invalid
-          ? "border-red-500/70 focus:border-red-400 focus:ring-red-400"
-          : "border-slate-700 focus:border-indigo-500 focus:ring-indigo-500")
-      }
-      title={
-        invalid
-          ? 'Enter a duration like "1h", "30m", "3 days"'
-          : "Filter by duration. Supports 1h, 30m, 3d, 1h30m, 2 hours…"
-      }
-    />
-  );
-}
-
-function BucketFilter({
-  selected,
-  onChange,
-}: {
-  selected: DurationBucket;
-  onChange: (next: DurationBucket) => void;
-}) {
-  const options: DurationBucket[] = [
-    "all",
-    "in_grace",
-    "short",
-    "long",
-    "ongoing",
-  ];
-  return (
-    <div
-      role="radiogroup"
-      aria-label="Filter breaches by duration"
-      className="flex flex-wrap gap-1.5"
-    >
-      {options.map((b) => {
-        const active = b === selected;
-        return (
-          <button
-            key={b}
-            role="radio"
-            type="button"
-            aria-checked={active}
-            onClick={() => !active && onChange(b)}
-            className={
-              "rounded-full px-2.5 py-0.5 text-[10px] uppercase tracking-wider font-medium transition-colors focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-indigo-400 " +
-              (active
-                ? "bg-slate-700 text-slate-200"
-                : "bg-slate-800/60 text-slate-400 hover:text-slate-200")
-            }
-          >
-            {BUCKET_LABEL[b]}
-          </button>
-        );
-      })}
     </div>
   );
 }
